+++ src/services/websocketserverservice.h
@@ -14,7 +14,19 @@
 
 #pragma once
 
+#include <QDateTime>
+#include <QHash>
+#include <QJsonArray>
+#include <QJsonObject>
+#include <QList>
 #include <QObject>
+#include <QPair>
+#include <QStringList>
+#include <QVector>
+
+class Bookmark;
+class QTcpServer;
+class QTcpSocket;
 
 class QWebSocketServer;
 class QWebSocket;
@@ -32,6 +44,14 @@ class WebSocketServerService : public QObject {
     static quint16 getSettingsPort();
 
     static quint16 getDefaultPort();
+
+    static bool isBookmarkSuggestionApiEnabled();
+
+    static quint16 getBookmarkSuggestionApiPort();
+
+    static quint16 getBookmarkSuggestionApiDefaultPort();
+
+    void refreshServers();
 
     void listen(quint16 port = 0);
 
@@ -62,8 +82,32 @@ class WebSocketServerService : public QObject {
     QWebSocketServer *m_pWebSocketServer;
     QList<QWebSocket *> m_clients;
     quint16 m_port{};
+    QTcpServer *m_pHttpServer;
+    quint16 m_httpPort{};
+    QHash<QString, QString> m_suggestionCache;
+    QVector<Bookmark> m_cachedBookmarks;
+    QDateTime m_bookmarkCacheTimestamp;
 
     static QString getBookmarksJsonText(bool hideCurrent = false);
+
+    QVector<Bookmark> getBookmarksForSuggestions();
+
+    QString homepageSuggestionResponse(const QString &query, int limit);
+
+    static QPair<QString, QHash<QString, QString>> parseHttpRequestLineAndQuery(
+        const QString &request);
+
+    static QString httpResponse(
+        int statusCode, const QByteArray &body, const QString &statusText = QStringLiteral("OK"),
+        const QString &contentType = QStringLiteral("application/json; charset=utf-8"));
+
+    void startSuggestionHttpServer();
+
+    void stopSuggestionHttpServer();
+
+    void handleHttpConnection();
+
+    void processHttpRequest(QTcpSocket *socket, const QString &requestText);
 
     static QString getCommandSnippetsJsonText();